// src/enums/constants.ts
export enum Constants {
  DOCKS_NUMBER = 4,
  DOCK_HEIGHT = window.innerHeight / 4.5,
  DOCK_WIDTH = 100,
  DOCK_GAP = 15,
  DOCK_COLOR = 0x00ffcc33,
  DOCK_LINE = 5,

  SHIP_WIDTH = 100,
  SHIP_HEIGHT = 40,
  SHIP_COLOR_RED = 0x00ff3300,
  SHIP_COLOR_GREEN = 0x0000ff33,
  SHIP_COLOR_EMPTY = 0x00336699,

  GATE_X = 400,
  GATE_Y = window.innerHeight / 2 - 50,

  SHIP_INTERVAL_TIME = 8000,
  STAY_IN_PORT_TIME = 5000,
  ANIMATION_TIME = 2000,

  QUEUE_X = 450,
  RED_QUEUE_Y = window.innerHeight / 2 - 150,
  GREEN_QUEUE_Y = window.innerHeight / 2 + 150,
}